+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed race condition with crafting tasks
 - Fixed pressing escape in crafting settings GUI not going back to grid GUI
 - Fixed losing autoselection in Grid when clicking on slot with autoselection mode
+- Fixed being able to pick up from pattern result slot
 - Added a max crafting quantity per request cap (hardcoded to 100)
 
 **Features**